+++ src/settings.ts
@@ -6,6 +6,8 @@ import { Zone } from "./zone";
 import { SystemZone } from "./zones/systemZone";
 import { ZoneLike } from "./types/zone";
 import { validateWeekSettings } from "./impl/util";
+import { DateTime } from "./datetime";
+import { resetDigitRegexCache } from "./impl/digits";
 
 let now = (): number => Date.now(),
     defaultZone: ZoneLike | null = "system",
@@ -140,9 +142,11 @@ export class Settings {
     }
 
     /**
-     * Set the cutoff year after which a string encoding a year as two digits is interpreted to occur in the current century.
-     * @example Settings.twoDigitCutoffYear = 0 // cut-off year is 0, so all 'yy' are interpreted as current century
-     * @example Settings.twoDigitCutoffYear = 50 // '49' -> 1949; '50' -> 2050
+     * Set the cutoff year for whether a 2-digit year string is interpreted in the current or previous century. Numbers higher than the cutoff will be considered to mean 19xx and numbers lower or equal to the cutoff will be considered 20xx.
+     * @type {number}
+     * @example Settings.twoDigitCutoffYear = 0 // all 'yy' are interpreted as 20th century
+     * @example Settings.twoDigitCutoffYear = 99 // all 'yy' are interpreted as 21st century
+     * @example Settings.twoDigitCutoffYear = 50 // '49' -> 2049; '50' -> 1950
      * @example Settings.twoDigitCutoffYear = 1950 // interpreted as 50
      * @example Settings.twoDigitCutoffYear = 2050 // ALSO interpreted as 50
      */
@@ -151,7 +155,8 @@ export class Settings {
     }
 
     /**
-     * Get the cutoff year after which a string encoding a year as two digits is interpreted to occur in the current century.
+     * Get the cutoff year for whether a 2-digit year string is interpreted in the current or previous century. Numbers higher than the cutoff will be considered to mean 19xx and numbers lower or equal to the cutoff will be considered 20xx.
+     * @type {number}
      */
     static get twoDigitCutoffYear(): number {
         return twoDigitCutoffYear;
@@ -165,6 +170,8 @@ export class Settings {
     static resetCaches(): void {
         Locale.resetCache();
         IANAZone.resetCache();
+        DateTime.resetCache();
+        resetDigitRegexCache();
     }
 
 }