+++ src/zones/fixedOffsetZone.ts
@@ -14,38 +14,63 @@ export class FixedOffsetZone extends Zone {
      * Get a singleton instance of UTC
      * @return {FixedOffsetZone}
      */
-    static get utcInstance() {
+    static get utcInstance(): FixedOffsetZone {
         if (singleton === null) {
             singleton = new FixedOffsetZone(0);
         }
         return singleton;
     }
 
-    /** @override **/
-    get isValid() {
-        return true;
-    }
-
-    get ianaName() {
+    /**
+     * The IANA name of this zone, i.e. `Etc/UTC` or `Etc/GMT+/-nn`
+     *
+     * @override
+     * @type {string}
+     */
+    get ianaName(): string {
         return this._fixed === 0
             ? "Etc/UTC"
             : `Etc/GMT${formatOffset(-this._fixed, "narrow")}`;
 
     }
 
-    /** @override **/
-    get name() {
+    /**
+     * Returns whether the offset is known to be fixed for the whole year:
+     * Always returns true for all fixed offset zones.
+     * @override
+     * @type {boolean}
+     */
+    get isUniversal(): boolean {
+        return true;
+    }
+
+    /**
+     * Return whether this Zone is valid:
+     * All fixed offset zones are valid.
+     * @override
+     * @type {boolean}
+     */
+    get isValid(): true {
+        return true;
+    }
+
+    /**
+     * The name of this zone.
+     * All fixed zones' names always start with "UTC" (plus optional offset)
+     * @override
+     * @type {string}
+     */
+    get name(): string {
         return this._fixed === 0 ? "UTC" : `UTC${formatOffset(this._fixed, "narrow")}`;
     }
 
-    /** @override **/
-    get type() {
+    /**
+     * The type of zone. `fixed` for all instances of `FixedOffsetZone`.
+     * @override
+     * @type {string}
+     */
+    get type(): "fixed" {
         return "fixed";
-    }
-
-    /** @override **/
-    get isUniversal() {
-        return true;
     }
 
     private readonly _fixed: number;
@@ -62,7 +87,7 @@ export class FixedOffsetZone extends Zone {
      * @param {number} offset - The offset in minutes
      * @return {FixedOffsetZone}
      */
-    static instance(offset: number) {
+    static instance(offset: number): FixedOffsetZone {
         return offset === 0 ? FixedOffsetZone.utcInstance : new FixedOffsetZone(offset);
     }
 
@@ -74,7 +99,7 @@ export class FixedOffsetZone extends Zone {
      * @example FixedOffsetZone.parseSpecifier("UTC-6:00")
      * @return {FixedOffsetZone}
      */
-    static parseSpecifier(s: string) {
+    static parseSpecifier(s: string): FixedOffsetZone {
         if (s) {
             const r = s.match(/^utc(?:([+-]\d{1,2})(?::(\d{2}))?)?$/i);
             if (r) {
@@ -84,24 +109,47 @@ export class FixedOffsetZone extends Zone {
         return null;
     }
 
-    /** @override **/
-    offsetName() {
-        return this.name;
+    /**
+     * Return whether this Zone is equal to another zone (i.e. also fixed and same offset)
+     * @override
+     * @param {Zone} otherZone - the zone to compare
+     * @return {boolean}
+     */
+    equals(otherZone: FixedOffsetZone): boolean {
+        return otherZone.type === "fixed" && otherZone._fixed === this._fixed;
     }
 
-    /** @override **/
-    formatOffset(_ts_: number, format: ZoneOffsetFormat) {
+    /**
+     * Returns the offset's value as a string
+     * @override
+     * @param {number} ts - Epoch milliseconds for which to get the offset
+     * @param {string} format - What style of offset to return.
+     *                          Accepts 'narrow', 'short', or 'techie'. Returning '+6', '+06:00', or '+0600' respectively
+     * @return {string}
+     */
+    formatOffset(_ts_: number, format: ZoneOffsetFormat): string {
         return formatOffset(this._fixed, format);
     }
 
-    /** @override **/
-    offset() {
+    /**
+     * Return the offset in minutes for this zone at the specified timestamp.
+     *
+     * For fixed offset zones, this is constant and does not depend on a timestamp.
+     * @override
+     * @return {number}
+     */
+    offset(): number {
         return this._fixed;
     }
 
-    /** @override **/
-    equals(otherZone: FixedOffsetZone) {
-        return otherZone.type === "fixed" && otherZone._fixed === this._fixed;
+    /**
+     * Returns the offset's common name at the specified timestamp.
+     *
+     * For fixed offset zones this equals to the zone name.
+     * @override
+     */
+    offsetName(): string {
+        return this.name;
     }