+++ test/datetime/proto.test.ts
@@ -3,9 +3,8 @@
 import { DateTime } from "../../src";
 
 test("DateTime prototype properties should not throw when accessed", () => {
-  const d = DateTime.now();
-  expect(() =>
-    // @ts-ignore
-    Object.getOwnPropertyNames(d.__proto__).forEach(name => d.__proto__[name])
-  ).not.toThrow();
+    const d = DateTime.now();
+    expect(() =>
+        Object.getOwnPropertyNames(Object.getPrototypeOf(d)).forEach((name) => Object.getPrototypeOf(d)[name])
+    ).not.toThrow();
 });