+++ test/interval/proto.test.ts
@@ -1,7 +1,8 @@
 import { DateTime } from "../../src";
 
 test("Interval prototype properties should not throw when addressed", () => {
-  const i = DateTime.fromISO("2018-01-01").until(DateTime.fromISO("2018-01-02"));
-  const proto = Object.getPrototypeOf(i);
-  expect(() => Object.getOwnPropertyNames(proto).forEach(name => proto[name])).not.toThrow();
+    const i = DateTime.fromISO("2018-01-01").until(DateTime.fromISO("2018-01-02"));
+    expect(() =>
+        Object.getOwnPropertyNames(Object.getPrototypeOf(i)).forEach((name) => Object.getPrototypeOf(i)[name])
+    ).not.toThrow();
 });